jenkins.py: check path exists before os.listdir

+++ jenkins.py
@@ -123,9 +123,11 @@ def run():
 
   if parsed_args.clear_except_download:
     base_dir = os.path.join(hunter_root, '_Base')
-    for filename in os.listdir(base_dir):
-      if filename != 'Download':
-        shutil.rmtree(os.path.join(base_dir, filename))
+    if os.path.exists(base_dir):
+      print('Clearing directory: {}'.format(base_dir))
+      for filename in os.listdir(base_dir):
+        if filename != 'Download':
+          shutil.rmtree(os.path.join(base_dir, filename))
 
   build_script = 'build.py'
   if os.name == 'nt':
